Updated config file and more JS refactoring.
